fastapi/requirements.txt
tensorflow == 2.9.1
pillow==9.0.1
opencv-python
torch
torchvision
fastapi
uvicorn
python-multipart
matplotlib
numpy
aiofiles